Refactored tests to not be packages
