+++ tests/wrapper/test_solc_wrapper.py
@@ -10,21 +10,14 @@ from solc.wrapper import (
     solc_wrapper,
 )
 
-from ..test_utils import (
-    skipif_no_standard_json,
-    checks_solc_version,
-)
-
 
 @pytest.fixture()
-@checks_solc_version
-def FOO_SOURCE():
+def FOO_SOURCE(supported_solc_version):
     return b"pragma solidity ^0.4.0;\ncontract Foo { function Foo() {} }"
 
 
 @pytest.fixture()
-@checks_solc_version
-def BAR_SOURCE():
+def BAR_SOURCE(supported_solc_version):
     return b"pragma solidity ^0.4.0;\ncontract Bar { function Bar() {} }"
 
 
@@ -76,7 +69,7 @@ def test_providing_multiple_source_files(contracts_dir, FOO_SOURCE, BAR_SOURCE):
     assert not err or err == 'Warning: This is a pre-release compiler version, please do not use it in production.\n'
 
 
-@skipif_no_standard_json
+@pytest.mark.requires_standard_json
 def test_providing_standard_json_input(FOO_SOURCE, BAR_SOURCE):
     stdin = json.dumps({
         "language": "Solidity",
